// src/app/models/notification.dao.ts
/* eslint-disable @typescript-eslint/naming-convention */
export enum NotificationDAO {
  ID = 'id',
  TITLE = 'title',
  PARAGRAPH = 'paragraph',
  CREATED_AT = 'createdAt',
  UNREAD = 'unread',
  PRODUCTS = 'products',
  LINK = 'link',
  ICON = 'icon',
  BUTTON_TEXT = 'button_text',
  BUTTON_ACTION = 'button_ga_action',
  TABLE_NAME = 'notification'
};